adding a bash script for mostly vanilla kernel
This is a bash script to automate building the stable series of 4.4.x kernel on Slackware. It should be trivial to adapt its use for the other series of kernels as well. It downloads the Linux kernel source, applies a GCC optimization patch, compiles it, and then copies the files to /boot.